Fix -Wfloat-conversion

+++ src/core/grid.cpp
@@ -51,7 +51,7 @@ int map_position_node_array(const Utils::Vector3d &pos) {
 
   Utils::Vector3i im;
   for (int i = 0; i < 3; i++) {
-    im[i] = std::floor(f_pos[i] / local_geo.length()[i]);
+    im[i] = static_cast<int>(std::floor(f_pos[i] / local_geo.length()[i]));
     im[i] = boost::algorithm::clamp(im[i], 0, node_grid[i] - 1);
   }
